azure-certifications-study-guide/AI-projects/project-10-anomaly-detection/src/requirements.txt
azure-ai-anomalydetector==3.0.0b6
azure-eventhub==5.11.0
azure-functions==1.17.0
azure-cosmos==4.5.1
python-dotenv==1.0.0
